working on delete cart

+++ src/components/cartIncart.js
@@ -1,31 +1,34 @@
 import React from "react";
 import { Container, Row, Col } from "react-bootstrap";
-import {Button} from 'react-bootstrap'
+import { Button } from "react-bootstrap";
 import { Card } from "@material-ui/core";
 import TableInCart from "./TableIncart";
 import { BASE_URL } from "../constants/BaseURL";
-import DeleteOutlineIcon from '@material-ui/icons/DeleteOutline';
-import '../assets/css/cartIncart.css'
+import DeleteOutlineIcon from "@material-ui/icons/DeleteOutline";
+import "../assets/css/cartIncart.css";
+import swal from "sweetalert";
+import axios from "axios";
 
 export default class CartInCart extends React.Component {
   constructor() {
     super();
     this.state = {
       quantity: "",
-      countOfItems:0
+      countOfItems: 0,
+      productsInCart: "",
     };
   }
   componentWillMount() {
     const x = JSON.parse(localStorage.getItem("cartProducts"));
-    console.log("length of cart==",x.length);
-    this.setState({ quantity: 1 ,countOfItems:x.length});
+    console.log("length of cart==", x.length);
+    this.setState({ quantity: 1, countOfItems: x.length, productsInCart: x });
   }
   incrementBtnHandler = (data, id) => {
     console.log("data in incHandler", data, id);
     data = data + 1;
     console.log("leng==", localStorage.length);
-    const temData = JSON.parse(localStorage.getItem("cartProducts"));
-
+    // const temData = JSON.parse(localStorage.getItem("cartProducts"));
+    const temData = this.state.productsInCart;
     temData.map((val, i) => {
       if (i === 0) {
         if (val[0].productDetail.id === id) {
@@ -40,14 +43,14 @@ export default class CartInCart extends React.Component {
     });
     localStorage.setItem("cartProducts", JSON.stringify(temData));
     console.log("leng==", localStorage.length);
-    this.setState({ quantity: data });
+    this.setState({ quantity: data, productsInCart: temData });
   };
-  decrementHandler = (data,id) => {
+  decrementHandler = (data, id) => {
     console.log("data in incHandler", data, id);
     data = data - 1;
     console.log("leng==", localStorage.length);
-    const temData = JSON.parse(localStorage.getItem("cartProducts"));
-
+    // const temData = JSON.parse(localStorage.getItem("cartProducts"));
+    const temData = this.state.productsInCart;
     temData.map((val, i) => {
       if (i === 0) {
         if (val[0].productDetail.id === id) {
@@ -62,18 +65,49 @@ export default class CartInCart extends React.Component {
     });
     localStorage.setItem("cartProducts", JSON.stringify(temData));
     console.log("leng==", localStorage.length);
-    this.setState({ quantity: data });
+    this.setState({ quantity: data, productsInCart: temData });
   };
-  deletehandler = (i) =>{
-    const temData = JSON.parse(localStorage.getItem("cartProducts"));
-    temData.splice(i,1);
-    localStorage.setItem("cartProducts", JSON.stringify(temData));
-  }
+  deletehandler = (i) => {
+    let token = localStorage.getItem("token");
+    swal({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    }).then((willDelete) => {
+      // ()
+      if (willDelete) {
+        console.log("in 1st if");
+        const temData = this.state.productsInCart;
+        temData.splice(i, 1);
+        localStorage.setItem("cartProducts", JSON.stringify(temData));
+        console.log("afetr splice tempdata ==",temData);
+        // if (token) {
+        //   console.log("in 2nd if of delete");
+        //   axios.delete(BASE_URL + `deleteCustomerCart/${i}`, {
+        //     headers: { Authorization: "Bearer " + token },
+        //   });
+        // }
+        swal({
+          icon: "success",
+          title: "Done",
+          text: "Item has been removed from cart.",
+          showConfirmButton: false,
+          timer: 2000,
+        });
+        console.log("before setState productdata");
+        this.setState({ productsInCart: temData });
+      }
+    });
+  };
   render() {
-    const x = JSON.parse(localStorage.getItem("cartProducts"));
-    console.log("in cart.js,,,data==", x);
-    console.log("length", x.length);
-    
+    // const x = JSON.parse(localStorage.getItem("cartProducts"));
+    // console.log("in cart.js,,,data==", x);
+    // console.log("length", x.length);
+    console.log("this.state.productCart==",this.state.productsInCart);
+    const x = this.state.productsInCart;
+    console.log("x==",x);
     return (
       <React.Fragment>
         <Container fluid={true}>
@@ -170,8 +204,15 @@ export default class CartInCart extends React.Component {
                                   </div>
                                 </td>
                                 <td>{data.price}</td>
-                                <td>{data.price*val.quantity}</td>
-                                <td><button style={{border:"none"}} onClick={()=>this.deletehandler(i)}><DeleteOutlineIcon /></button></td>
+                                <td>{data.price * val.quantity}</td>
+                                <td>
+                                  <button
+                                    style={{ border: "none" }}
+                                    onClick={() => this.deletehandler(i)}
+                                  >
+                                    <DeleteOutlineIcon />
+                                  </button>
+                                </td>
                               </tr>
                             );
                           } else {
@@ -216,8 +257,12 @@ export default class CartInCart extends React.Component {
                                         border: "none",
                                         background: "red",
                                       }}
-                                      onClick={()=>this.decrementHandler(val[0].quantity,
-                                        data1.id)}
+                                      onClick={() =>
+                                        this.decrementHandler(
+                                          val[0].quantity,
+                                          data1.id
+                                        )
+                                      }
                                     >
                                       -
                                     </button>
@@ -246,8 +291,15 @@ export default class CartInCart extends React.Component {
                                   </div>
                                 </td>
                                 <td>{data1.price}</td>
-                                <td>{data1.price*val[0].quantity }</td>
-                                <td><button style={{border:"none"}} onClick={()=>this.deletehandler(i)}><DeleteOutlineIcon /></button></td>
+                                <td>{data1.price * val[0].quantity}</td>
+                                <td>
+                                  <button
+                                    style={{ border: "none" }}
+                                    onClick={() => this.deletehandler(i)}
+                                  >
+                                    <DeleteOutlineIcon />
+                                  </button>
+                                </td>
                               </tr>
                             );
                           }
@@ -258,18 +310,26 @@ export default class CartInCart extends React.Component {
               </Card>
             </Col>
             <Col xs={4}>
-                <React.Fragment>
-                    <Card>
-                            <h3>Review Order</h3>
-                            <ul>
-                                <li className="revieworderItems"><span>Subtotal</span><span></span></li>
-                                <li className="revieworderItems"><span>GST(5%)</span><span></span></li>
-                                <li className="revieworderItems"><span>Order Total</span><span></span></li>
-
-                            </ul>
-                            <Button variant="primary">Proceed to buy</Button>
-                    </Card>
-                </React.Fragment>
+              <React.Fragment>
+                <Card>
+                  <h3>Review Order</h3>
+                  <ul>
+                    <li className="revieworderItems">
+                      <span>Subtotal</span>
+                      <span></span>
+                    </li>
+                    <li className="revieworderItems">
+                      <span>GST(5%)</span>
+                      <span></span>
+                    </li>
+                    <li className="revieworderItems">
+                      <span>Order Total</span>
+                      <span></span>
+                    </li>
+                  </ul>
+                  <Button variant="primary">Proceed to buy</Button>
+                </Card>
+              </React.Fragment>
             </Col>
           </Row>
         </Container>
